Add libs for CheckVersion

+++ src/erp.au3
@@ -14,18 +14,55 @@
 #EndRegion ;**** Directives created by AutoIt3Wrapper_GUI ****
 
 #include-once
-#include "SystemRestore.au3""
+#include <GUIConstantsEx.au3>
+#include <WindowsConstants.au3>
+#include <InetConstants.au3>
+#include <MsgBoxConstants.au3>
+#include <Date.au3>
+#include <WinAPIFiles.au3>
+#include "SystemRestore.au3"
+#include "HTTP.au3"
+#include "_SelfUpdate.au3"
 
 Global $sToolVersion = '0.1'
+Global $sToolReportFile = "easyrestorepoint.txt"
 
-Func QuitTool($bAutoDelete = False)
-	If $bAutoDelete Then
+Global $bDevVersion = False
+
+If $bDevVersion = True Then
+	AutoItSetOption("MustDeclareVars", 1)
+EndIf
+
+If FileExists(@DesktopDir & "\" & $sToolReportFile) Then
+	FileDelete(@DesktopDir & "\" & $sToolReportFile)
+EndIf
+
+Func LogMessage($message)
+	Dim $sToolReportFile
+
+	FileWrite(@DesktopDir & "\" & $sToolReportFile, $message & @CRLF)
+EndFunc   ;==>LogMessage
+
+If Not IsAdmin() Then
+	MsgBox(16, "Fail", "Admin required")
+	QuitTool()
+EndIf
+
+Func QuitTool($bAutoDelete = False, $bOpen = False)
+	Dim $sToolReportFile
+	Dim $bDevVersion
+
+	If $bOpen Then
+		Run("notepad.exe " & @DesktopDir & "\" & $sToolReportFile)
+	EndIf
+
+	If $bAutoDelete = True And $bDevVersion = False Then
 		Run(@ComSpec & ' /c timeout 3 && del /F /Q "' & @ScriptFullPath & '"', @TempDir, @SW_HIDE)
 		FileDelete(@ScriptFullPath)
 	EndIf
 
 	Exit
-EndFunc   ;==>QuitKprm
+EndFunc   ;==>QuitTool
 
 Func _IsInternetConnected()
 	Local $aReturn = DllCall('connect.dll', 'long', 'IsInternetConnected')
@@ -43,7 +80,7 @@ Func CheckVersion()
 	If _IsInternetConnected() = False Then Return
 
 	Local $aCurrentVersion = StringSplit($sToolVersion, ".")
-	Local $sVersion = _HTTP_Get("https://toolslib.net/api/softwares/951/version")
+	Local $sVersion = _HTTP_Get("https://toolslib.net/api/softwares/979/version")
 	Local $aVersion = StringSplit($sVersion, ".")
 	Local $bNeedsUpdate = False
 
@@ -60,15 +97,15 @@ Func CheckVersion()
 
 		If @error Then
 			MsgBox($MB_SYSTEMMODAL, "Easy Restore Point - Updater", "The script must be a compiled exe to work correctly or the update file must exist.")
-			QuitTool(True)
+			QuitTool()
 		EndIf
 	EndIf
-EndFunc   ;==>CheckVersionOfKpRm
+EndFunc   ;==>CheckVersion
 
 Func DownloadLatest()
 	ProgressOn("Easy Restore Point - Updater", "Downloading..", "0%")
 	; Download from the following URL.
-	Local $sURL = "https://download.toolslib.net/download/direct/951/latest"
+	Local $sURL = "https://download.toolslib.net/download/direct/979/latest"
 	; Save the downloaded file to the temporary folder.
 	Local $sFilePath = _WinAPI_GetTempFileName(@TempDir)
 	; Remote file size
@@ -136,6 +173,8 @@ EndFunc   ;==>ClearDayRestorePoint
 Func ShowCurrentRestorePoint()
 	Local Const $aRP = _SR_EnumRestorePoints()
 
+	LogMessage(@CRLF & "- Display All System Restore Point -" & @CRLF)
+
 	If $aRP[0][0] = 0 Then
 		LogMessage("No System Restore point found")
 		Return
@@ -153,8 +192,8 @@ Func CreateSystemRestorePoint()
 	Return @error
 EndFunc   ;==>CreateSystemRestorePoint
 
-
 Func CreateRestorePoint()
+	ProgressOn("Easy Restore Point", "Run", "0%")
 	Local $iSR_Enabled = _SR_Enable()
 
 	If $iSR_Enabled = 0 Then
@@ -162,9 +201,51 @@ Func CreateRestorePoint()
 		$iSR_Enabled = _SR_Enable()
 	EndIf
 
+	ProgressSet(25, 25 & "%")
+
 	ClearDayRestorePoint()
 	Sleep(1500)
+	ProgressSet(50, 50 & "%")
+
 	CreateSystemRestorePoint()
 	Sleep(1500)
+	ProgressSet(75, 75 & "%")
+
 	ShowCurrentRestorePoint()
+	ProgressSet(100, 100 & "%")
+
+	ProgressOff()
 EndFunc   ;==>CreateRestorePoint
+
+Func GetHumanVersion()
+	Switch @OSVersion
+		Case "WIN_VISTA"
+			Return "Windows Vista"
+		Case "WIN_7"
+			Return "Windows 7"
+		Case "WIN_8"
+			Return "Windows 8"
+		Case "WIN_81"
+			Return "Windows 8.1"
+		Case "WIN_10"
+			Return "Windows 10"
+		Case Else
+			Return "Unsupported OS"
+	EndSwitch
+EndFunc   ;==>GetHumanVersion
+
+If $bDevVersion = False Then
+	CheckVersion()
+EndIf
+
+LogMessage("# Run at " & _Now())
+LogMessage("# Easy Restore Point (Kernel-panik) version " & $sToolVersion)
+LogMessage("# Website https://kernel-panik.me/")
+LogMessage("# Run by " & @UserName & " from " & @WorkingDir)
+LogMessage("# Computer Name: " & @ComputerName)
+LogMessage("# OS: " & GetHumanVersion() & " " & @OSArch & " (" & @OSBuild & ") " & @OSServicePack)
+
+CreateRestorePoint()
+
+QuitTool(True, True)
+
